+++ resources/views/admin/user/edit.blade.php
@@ -45,6 +45,11 @@
 
             <x-create.zweite-stufe :checked="$user->two_factor_required" />
 
+            {{-- Nur beim Bearbeiten: Ein Zugang, der gesperrt angelegt wird,
+                 ergaebe keinen Sinn. --}}
+            <x-create.gesperrt :checked="$user->istDeaktiviert()" :seit="$user->deactivated_at" />
+            <x-input.error :messages="$errors->get('deactivated')" class="mt-2" />
+
         </x-create.main>
 
         @if ($errors->has('einladung'))